+++ server/repository/EntryRepository.go
@@ -55,7 +55,7 @@ func (repo *EntryRepository) Update(
 }
 
 func (repo *EntryRepository) Delete(id int) (int, error) {
-	entry := structs.Entry{Base: structs.Base{ID: uint(id)}}
+	entry := structs.Entry{Base: structs.Base{ID: id}}
 	err := application.Db.Delete(&entry).Error
 	return int(entry.ID), err
 }
